--- src/main/java/com/clipclap/rego/crawler/airLineDTO.java
@@ -1,17 +1,0 @@
-package com.clipclap.rego.crawler;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class airLineDTO {
-    private String airline;
-    private String departureTime;
-    private String departureAirport;
-    private String arrivalTime;
-    private String arrivalAirport;
-    private String duration;
-}
