--- src/main/resources/templates/crawl/crawl.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Airlines</title>
-</head>
-<body>
-<table>
-    <thead>
-    <tr>
-        <th>Airline</th>
-        <th>Departure Time</th>
-        <th>Departure Airport</th>
-        <th>Arrival Time</th>
-        <th>Arrival Airport</th>
-        <th>Duration</th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr th:each="airline : ${airlines}">
-        <td th:text="${airline.airline}"></td>
-        <td th:text="${airline.departureTime}"></td>
-        <td th:text="${airline.departureAirport}"></td>
-        <td th:text="${airline.arrivalTime}"></td>
-        <td th:text="${airline.arrivalAirport}"></td>
-        <td th:text="${airline.duration}"></td>
-    </tr>
-    </tbody>
-</table>
-</body>
-</html>
