+++ src/main/resources/templates/crawl/roundCrawl.html
@@ -2,37 +2,92 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <title>Airlines</title>
+    <link href="/css/flightSearchResult.css" rel="stylesheet" type="text/css">
 </head>
 <body>
-<table border="1">
-    <thead>
-    <tr>
-        <th>Airline</th>
-        <th>Departure</th>
-        <th>Arrival</th>
-        <th>Duration</th>
-        <th>Price</th>
-    </tr>
-    </thead>
-    <tbody>
-    <th:block th:each="flight : ${flights}">
-        <!-- 첫 번째 루트 (편도) -->
-        <tr>
-            <td><img th:src="${flight.routes[0].airlineImg}" alt="${flight.routes[0].airlineName}" width="50"> <span th:text="${flight.routes[0].airlineName}"></span></td>
-            <td th:text="${flight.routes[0].departureTime} + ' (' + ${flight.routes[0].departureAirport} + ')'"></td>
-            <td th:text="${flight.routes[0].arrivalTime} + ' (' + ${flight.routes[0].arrivalAirport} + ')'"></td>
-            <td th:text="${flight.routes[0].duration}"></td>
-            <td rowspan="2" th:text="${flight.price}"></td>
-        </tr>
-        <!-- 두 번째 루트 (복귀) -->
-        <tr>
-            <td><img th:src="${flight.routes[1].airlineImg}" alt="${flight.routes[1].airlineName}" width="50"> <span th:text="${flight.routes[1].airlineName}"></span></td>
-            <td th:text="${flight.routes[1].departureTime} + ' (' + ${flight.routes[1].departureAirport} + ')'"></td>
-            <td th:text="${flight.routes[1].arrivalTime} + ' (' + ${flight.routes[1].arrivalAirport} + ')'"></td>
-            <td th:text="${flight.routes[1].duration}"></td>
-        </tr>
-    </th:block>
-    </tbody>
-</table>
+<div class="content">
+    <div class="nav"
+         style=" box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.2); padding-bottom: 15px; background-color: white; width: 1700px; height: 73px;  text-align: center; position:fixed; top: 0px;">
+
+        <ul class="" style="list-style-type:none;">
+            <li class="">
+                <!-- 타임리프의 security이용
+                sec:authorize="isAnonymous()은 로그인하지않는 경우->로그인 출력
+                sec:authorize="isAuthenticated()은 로그인한 경유->로그아웃 출력
+                -->
+                <a class="logo" th:href="@{/}"><img src="/img/logo.png"
+                                                    style="float: left; width: 63px; height: 58px; margin-left: 10px; background-repeat: no-repeat;"></a>
+
+
+
+                <div class="nav-left"
+                     style="float:left;   margin-left: 470px; margin-top: 19px;  font-size: 20px;  font-family: 'Roboto', sans-serif; font-weight: 500;">
+                    <a class="" th:href="@{/}"
+                       style="margin-right: 50px; color:#4F54B4;">HOME</a>
+                    <a class="" th:href="@{/}"
+                       style="margin-right: 50px; color:#4F54B4;">AIRPORT</a>
+                    <a class="" th:href="@{/}"
+                       style="margin-right: 50px; color:#4F54B4;">START PLAN</a>
+                    <a class="" th:href="@{/}"
+                       style="color:#4F54B4;">RECOMMEND</a>
+                </div>
+                <div class="nav-right"
+                     style="float:right;    margin-right: 35px; margin-top: 27px;  font-size: 12px;  font-family: 'Roboto', sans-serif; font-weight: 600;">
+                    <a class="" th:href="@{/}"
+                       style="margin-right: 20px; color:#4F54B4;">고객센터</a>
+                    <a class="loginBtn" sec:authorize="isAnonymous()" th:href="@{/login}"
+                       style="margin-right: 20px; color:#4F54B4;">로그인</a>
+                    <a class="" sec:authorize="isAuthenticated()" th:href="@{/}"
+                       style="margin-right: 20px; color:#4F54B4;">마이페이지</a>
+                    <a class="" sec:authorize="isAuthenticated()" th:href="@{/}"
+                       style="margin-right: 20px; color:#4F54B4;">로그아웃</a>
+                    <a class="" sec:authorize="hasRole('ADMIN')" th:href="@{/}"
+                       style="color:#4F54B4;">ADMIN</a>
+                </div>
+            </li>
+
+        </ul>
+        <div>
+
+        </div>
+    </div>
+    <!-- nav끝   -->
+    <div class="navEmptyBox">
+
+    </div>
+    <div class="resultBox">
+        <table border="1">
+            <thead>
+            <tr>
+                <th>Airline</th>
+                <th>Departure</th>
+                <th>Arrival</th>
+                <th>Duration</th>
+                <th>Price</th>
+            </tr>
+            </thead>
+            <tbody>
+            <th:block th:each="flight : ${flights}">
+                <!-- 첫 번째 루트 (편도) -->
+                <tr>
+                    <td><img th:src="${flight.routes[0].airlineImg}" alt="${flight.routes[0].airlineName}" width="50"> <span th:text="${flight.routes[0].airlineName}"></span></td>
+                    <td th:text="${flight.routes[0].departureTime} + ' (' + ${flight.routes[0].departureAirport} + ')'"></td>
+                    <td th:text="${flight.routes[0].arrivalTime} + ' (' + ${flight.routes[0].arrivalAirport} + ')'"></td>
+                    <td th:text="${flight.routes[0].duration}"></td>
+                    <td rowspan="2" th:text="${flight.price}"></td>
+                </tr>
+                <!-- 두 번째 루트 (복귀) -->
+                <tr>
+                    <td><img th:src="${flight.routes[1].airlineImg}" alt="${flight.routes[1].airlineName}" width="50"> <span th:text="${flight.routes[1].airlineName}"></span></td>
+                    <td th:text="${flight.routes[1].departureTime} + ' (' + ${flight.routes[1].departureAirport} + ')'"></td>
+                    <td th:text="${flight.routes[1].arrivalTime} + ' (' + ${flight.routes[1].arrivalAirport} + ')'"></td>
+                    <td th:text="${flight.routes[1].duration}"></td>
+                </tr>
+            </th:block>
+            </tbody>
+        </table>
+    </div>
+</div>
+
 </body>
 </html>
